chore: post 분리
post를 course에서 분리함
