+++ src/main/java/com/example/p2k/course/Course.java
@@ -1,15 +1,10 @@
 package com.example.p2k.course;
 
-import com.example.p2k.course.courseuser.CourseUser;
 import com.example.p2k.user.BaseTimeEntity;
-import com.example.p2k.user.User;
 import jakarta.persistence.*;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-
-import java.util.ArrayList;
-import java.util.List;
 
 @Getter
 @NoArgsConstructor
