+++ src/main/java/com/example/p2k/post/Category.java
@@ -1,4 +1,4 @@
-package com.example.p2k.course.post;
+package com.example.p2k.post;
 
 public enum Category {
     NOTICE, QUESTION, FREE
